requirements.txt
speechrecognition==3.10.4
pyttsx3==2.90
pypiwin32==223 ; platform_system == "Windows"
openai>=1.30.0
python-dotenv>=1.0.1
requests>=2.32.3
feedparser>=6.0.11
wikipedia>=1.4.0
sgmllib3k>=1.0.0
pyinstaller>=6.11.0
Pillow>=10.4.0

